feat: move subscriptions to givings
