+++ app/controllers/api/v1/subscriptions/subscriptions_controller.rb
@@ -2,13 +2,11 @@ module Api
   module V1
     module Subscriptions
       class SubscriptionsController < ApplicationController
-        include ::Subscriptions
-
         def unsubscribe
           subscription_id = ::Jwt::Decoder.decode(token: params[:token]).first['subscription_id']
           return head :unauthorized unless subscription_id
 
-          command = ::Subscriptions::CancelSubscription.call(subscription_id:)
+          command = ::Givings::Subscriptions::CancelSubscription.call(subscription_id:)
 
           if command.success?
             render json: SubscriptionBlueprint.render(command.result.subscription), status: :ok
