+++ app/controllers/api/v1/users/subscriptions_controller.rb
@@ -2,7 +2,6 @@ module Api
   module V1
     module Users
       class SubscriptionsController < ApplicationController
-        include ::Subscriptions
         def index
           ids = current_user.customers.pluck(:id)
           @subscriptions = Subscription.where(payer_id: ids, status: :active)
@@ -11,7 +10,7 @@ module Api
         end
 
         def send_cancel_subscription_email
-          command = ::Subscriptions::SendCancelSubscriptionEmail.new(subscription:).call
+          command = ::Givings::Subscriptions::SendCancelSubscriptionEmail.new(subscription:).call
 
           if command.success?
             render json: { message: 'Email sent' }, status: :ok
