+++ spec/requests/api/v1/subscriptions/subscriptions_spec.rb
@@ -3,11 +3,11 @@ require 'rails_helper'
 RSpec.describe 'Api::V1::Subscriptions::Subscriptions', type: :request do
   let(:subscription) { create(:subscription) }
   let(:create_order_command_double) do
-    command_double(klass: Subscriptions::CancelSubscription)
+    command_double(klass: ::Givings::Subscriptions::CancelSubscription)
   end
 
   before do
-    allow(Subscriptions::CancelSubscription).to receive(:call).and_return(create_order_command_double)
+    allow(::Givings::Subscriptions::CancelSubscription).to receive(:call).and_return(create_order_command_double)
   end
 
   describe 'PUT /api/v1/subscriptions/cancel_subscription/:id' do
@@ -17,7 +17,7 @@ RSpec.describe 'Api::V1::Subscriptions::Subscriptions', type: :request do
       let(:token) { Jwt::Encoder.encode({ subscription_id: subscription.id }) }
 
       let(:create_order_command_double) do
-        command_double(klass: Subscriptions::CancelSubscription, success: true,
+        command_double(klass: ::Givings::Subscriptions::CancelSubscription, success: true,
                        result: OpenStruct.new(subscription:))
       end
 
@@ -38,7 +38,7 @@ RSpec.describe 'Api::V1::Subscriptions::Subscriptions', type: :request do
 
       let(:token) { Jwt::Encoder.encode({ subscription_id: subscription.id }) }
       let(:create_order_command_double) do
-        command_double(klass: Subscriptions::CancelSubscription, success: false)
+        command_double(klass: ::Givings::Subscriptions::CancelSubscription, success: false)
       end
 
       before do
